+++ server/schemas/Cheese.js
@@ -1,9 +1,9 @@
-const { Text, Float } = require("@keystonejs/fields");
+const { Text, Integer } = require("@keystonejs/fields");
 
 module.exports = {
   fields: {
     name: { type: Text },
-    price: { type: Float },
-    stock: { type: Float },
+    price: { type: Integer },
+    stock: { type: Integer },
   },
 };
